+++ test/flycheck-test.el
@@ -3833,6 +3833,13 @@ See https://github.com/flycheck/flycheck/issues/531 and Emacs bug #19206"))
         :checker emacs-lisp )))
 
 (flycheck-ert-def-checker-test (emacs-lisp emacs-lisp-checkdoc) emacs-lisp
+                               right-syntax-table
+  (flycheck-ert-should-syntax-check
+   "language/emacs-lisp/checkdoc-syntax-table-regression.el" 'emacs-lisp-mode
+   '(11 nil warning "All variables and subroutines might as well have a documentation string"
+        :checker emacs-lisp-checkdoc)))
+
+(flycheck-ert-def-checker-test (emacs-lisp emacs-lisp-checkdoc) emacs-lisp
                                checks-compressed-file
   (flycheck-ert-should-syntax-check
    "language/emacs-lisp/compressed.el.gz" 'emacs-lisp-mode
@@ -3844,7 +3851,7 @@ See https://github.com/flycheck/flycheck/issues/531 and Emacs bug #19206"))
                     "the function ‘dummy-package-foo’ is not known to be defined.")
         :checker emacs-lisp)))
 
-(flycheck-ert-def-checker-test emacs-lisp emacs-lisp sytnax-error
+(flycheck-ert-def-checker-test emacs-lisp emacs-lisp syntax-error
   (let ((flycheck-disabled-checkers '(emacs-lisp-checkdoc)))
     (flycheck-ert-should-syntax-check
      "language/emacs-lisp/syntax-error.el" 'emacs-lisp-mode