+++ CHANGES.md
@@ -26,6 +26,7 @@ now for questions and discussions about development. [GH-820]
   * Don’t change current buffer when closing the error message buffer [GH-648]
   * Never display error message buffer in current window [GH-822]
   * Work around a caching issue in Rubocop [GH-844]
+  * Fix checkdoc failure with some Emacs Lisp syntax [GH-833] [GH-845]
 
 ## 0.25.1 (Nov 16, 2015) ##